PHYS ITEMS FUNCTIONAL ACROSS ALL CLIENTS 🥳🥳
Theres only one issue with phys items pushing other phys items, this wont update their position on the server. I have other things to work on, and thats a fix I'm easily visualizing, but just dont have the time to work on because this is 1 mechanic out of many for this game, and we're on day 3-4.